+++ internal/gophermart/model/balances.go
@@ -13,10 +13,10 @@ type BalanceResponse struct {
 }
 
 type TransactionCreate struct {
-	Order       string `json:"order"`
-	Sum         int64  `json:"sum"`
-	UserID      string `json:"user_id"`
-	ProcessedAt string `json:"processed_at"`
+	Order       string  `json:"order"`
+	Sum         float64 `json:"sum"`
+	UserID      string  `json:"user_id"`
+	ProcessedAt string  `json:"processed_at"`
 }
 
 type TransactionCreateRequest struct {
